Initialization script with migrations for v1_0 and openAPI
This commit adds the foundations for the exposed schemas with
primary roles and privileges. Everything that is common to the
API such as request/response functions, pgjwt, and main roles are
available in the initialization script.

Everything that is specific to data (e.g users) and functions
(e.g signup) is provided with migrations. At the moment I didn't
find the right tool to auto-generate migration files based on
on schema diff. Planned for later.

Current SQL supports signup with rpc endpoint.